% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{hello}
\alias{hello}
\title{My first test function in my first package}
\usage{
hello(x)
}
\arguments{
\item{x}{Character, the name of the person}
}
\value{
The output string: \code{\link{print}}
}
\description{
My first test function in my first package
}
\examples{
hello("Niklas")
\dontrun{
hello("Steve")
}
}
